Add response to webserver
This allows speck.ly to keep sending requests, since it thinks its getting a valid response. This response needs work, the goal would be to mimic the greyhound webserver perfectly. For now the binary data that is responded is not similiar, and the visualized point cloud is strange. It appears to be single points only